fixed order id type

+++ internal/accrual/storage/migrations/000001_create_orders_table.up.sql
@@ -1,5 +1,5 @@
 create table if not exists orders(
-    id int unique not null,
+    id bigint unique not null,
     created_at timestamp not null default now(),
     status smallint not null,
     accrual decimal
